+++ geoimagenet_api/CHANGELOG.rst
@@ -1,11 +1,12 @@
 Changelog
 =========
 
-Unreleased
+1.4.2 (2020-07-17)
 ------------------
 
 Changes
 ~~~~~~~
+- Delete trace cache layer in gwc
 - Bugfix in how "batches/annotations" route gets Taxonomy ids